feat(hodo-list): Move archives from modal to page
