--- apps/hodo-list/src/components/calendar/constants.ts
@@ -1,9 +1,0 @@
-export const DaysOfWeek = {
-  SUNDAY: 0,
-  MONDAY: 1,
-  TUESDAY: 2,
-  WEDNESDAY: 3,
-  THURSDAY: 4,
-  FRIDAY: 5,
-  SATURDAY: 6,
-};
